removed client-connections interface
